// src/theme/colors.interface.ts
// 🔹 Color object uchun umumiy type
export interface IThemeColors {
  primary: string;
  pageBackground: string;
  Boxbackground: string;
  textPrimary: string;
  textSecondary: string;
  green: string;
  yellow: string;
  red: string;
  borderColor: string;

  primary08: string;
  primary06: string;
  primary04: string;
  primary02: string;

  textPrimary08: string;
  textPrimary06: string;
  textPrimary04: string;
  textPrimary02: string;

  textSecondary08: string;
  textSecondary06: string;
  textSecondary04: string;
  textSecondary02: string;

  green08: string;
  green06: string;
  green04: string;
  green02: string;

  yellow08: string;
  yellow06: string;
  yellow04: string;
  yellow02: string;

  red08: string;
  red06: string;
  red04: string;
  red02: string;

  Boxbackground08: string;
  Boxbackground06: string;
  Boxbackground04: string;
  Boxbackground02: string;

  borderColor08: string;
  borderColor06: string;
  borderColor04: string;
  borderColor02: string;
}